Fixed return type of function 'libifc_err_ioctlreq'. Added examples directory, currently with only one example. More to follow later.

+++ libifconfig.c
@@ -102,7 +102,7 @@ int libifc_err_errno(libifc_handle_t *h) {
         return h->error.errcode;
 }
 
-int libifc_err_ioctlreq(libifc_handle_t *h) {
+unsigned long libifc_err_ioctlreq(libifc_handle_t *h) {
         return h->error.ioctl_request;
 }
